Added collapsing to the nav menu
