+++ package.json
@@ -38,7 +38,7 @@
     "extract-text-webpack-plugin": "^2.0.0-rc.3",
     "file-loader": "^0.10.0",
     "jest": "^17.0.0",
-    "jquery": "^3.1.1",
+    "jquery": "^3.3.1",
     "less": "^2.7.2",
     "less-loader": "^2.2.3",
     "lodash": "^4.17.4",
